Split couplercalls and solver classes.
