+++ src/output/OpenWQ_output.hpp
@@ -28,42 +28,43 @@ class OpenWQ_output{
 
     public:
 
-    // main print function
-    int writeResults(
-        OpenWQ_json& OpenWQ_json,
-        OpenWQ_vars& OpenWQ_vars,
-        OpenWQ_hostModelconfig& OpenWQ_hostModelconfig,
-        OpenWQ_wqconfig& OpenWQ_wqconfig,
-        OpenWQ_solver& OpenWQ_solver,
-        time_t simtime);
+        // main print function
+        int writeResults(
+            OpenWQ_json& OpenWQ_json,
+            OpenWQ_vars& OpenWQ_vars,
+            OpenWQ_hostModelconfig& OpenWQ_hostModelconfig,
+            OpenWQ_wqconfig& OpenWQ_wqconfig,
+            OpenWQ_solver& OpenWQ_solver,
+            time_t simtime);
+        
+        // Print in console and log file
+        void ConsoleLog(
+            OpenWQ_wqconfig& OpenWQ_wqconfig,
+            std::string& msg_string,
+            bool print_console,
+            bool print_logFile);
+    private:
+        // Print output in CSV
+        int writeCSV(
+            OpenWQ_json& OpenWQ_json,
+            OpenWQ_hostModelconfig& OpenWQ_hostModelconfig,
+            OpenWQ_wqconfig& OpenWQ_wqconfig,
+            std::unique_ptr<arma::field<arma::field<arma::cube>>>& OpenWQ_var2print,
+            std::string& output_file_label,
+            std::string timestr,            // time step (in seconds)
+            int icmp);
+        
+    // Print output in HDF5
+        int writeHDF5(
+            OpenWQ_json& OpenWQ_json,
+            OpenWQ_hostModelconfig& OpenWQ_hostModelconfig,
+            OpenWQ_wqconfig& OpenWQ_wqconfig,
+            std::unique_ptr<arma::field<arma::field<arma::cube>>>& OpenWQ_var2print,
+            std::string& output_file_label,
+            std::string timestr,            // time step (in seconds)
+            int icmp);
 
-    // Print output in CSV
-    int writeCSV(
-        OpenWQ_json& OpenWQ_json,
-        OpenWQ_hostModelconfig& OpenWQ_hostModelconfig,
-        OpenWQ_wqconfig& OpenWQ_wqconfig,
-        std::unique_ptr<arma::field<arma::field<arma::cube>>>& OpenWQ_var2print,
-        std::string& output_file_label,
-        std::string timestr,            // time step (in seconds)
-        int icmp);
-    
-   // Print output in HDF5
-    int writeHDF5(
-        OpenWQ_json& OpenWQ_json,
-        OpenWQ_hostModelconfig& OpenWQ_hostModelconfig,
-        OpenWQ_wqconfig& OpenWQ_wqconfig,
-        std::unique_ptr<arma::field<arma::field<arma::cube>>>& OpenWQ_var2print,
-        std::string& output_file_label,
-        std::string timestr,            // time step (in seconds)
-        int icmp);
 
-
-    // Print in console and log file
-    void ConsoleLog(
-        OpenWQ_wqconfig& OpenWQ_wqconfig,
-        std::string& msg_string,
-        bool print_console,
-        bool print_logFile);
 
 };
